Refine activity views and preserve input capture precision in 7.0.1

+++ src/TimeLens.Api/Dtos/DashboardResponse.cs
@@ -68,7 +68,7 @@ public sealed record TopAppDto(
 
 public sealed record HeatmapEntryDto(
     string Date,
-    int Value
+    double Value
 );
 
 public sealed record CategoryEntryDto(
